chore: Update prettier configuration and package.json scripts
The `.prettierrc` file was modified to include the `endOfLine` option set to "auto". This change ensures consistent line endings in the codebase.

In the `package.json` file, the following scripts were updated:
- `start:dev` was renamed to `dev`
- `start:debug` was renamed to `debug`
- `start:prod` was renamed to `preview`

These changes improve the clarity and consistency of the script names.

The `license` field in `package.json` was also updated from "UNLICENSED" to "MIT" to reflect the correct license for the project.

The dependencies in `package.json` were updated to the latest versions of the `@nestjs` packages, including `@nestjs/config`, `@nestjs/jwt`, `@nestjs/mapped-types`, `@nestjs/mongoose`, `@nestjs/passport`, `@nestjs/platform-socket.io`, and `@nestjs/websockets`. Other dependencies such as `mongoose`, `passport`, `passport-jwt`, `class-transformer`, and `class-validator` were also updated to their latest versions.

The devDependencies in `package.json` were updated to the latest versions of the `@types` packages, including `@types/passport-jwt`.

These changes ensure that the project is using the latest versions of its dependencies and aligns with best practices for code formatting and script naming.